security and provider resolver improvement
* User now has an IsGuest property.  client system can use as desired.
It is meant to mean a user that was auto registered and invited.
Whereas a non guest signed up as a user directly.
* UserProvider now has a configuration for the number of hours a
password reset code is valid for.  Defaults to 1, the prior hardcoded
value for backwards compatibility.
* ProviderResolver errors caused by no providers configured for a type
will now include the type name in the error message.  Prior the message
was "No '' providers have been configured." (missing type name in single
quotes)